Rename: list 디렉토리를 dex로 변경
