+++ src/components/dex/PokemonListElement.tsx
